This adds a flush benchmark and an optimization to reduce memory usage.

```
❯ benchstat before.txt after.txt                      
name      old time/op    new time/op    delta
Flush-16     400ms ± 8%     249ms ±10%  -37.76%  (p=0.008 n=5+5)

name      old alloc/op   new alloc/op   delta
Flush-16    1.03GB ± 0%    0.20GB ± 1%  -80.12%  (p=0.008 n=5+5)

name      old allocs/op  new allocs/op  delta
Flush-16     8.26k ± 1%     3.43k ± 1%  -58.54%  (p=0.008 n=5+5)
```