@simonswine 2023-03-21
We also need to call `Flush()` to ensure the rowGroup boundaries are still respected.

So that one segment in becomes one row group in the final file. Otherwise the ordering is incorrect. Still significant win:

```
❯ benchstat before.txt after2.txt
name     old time/op    new time/op    delta
Flush-8     533ms ±13%     294ms ± 6%  -44.84%  (p=0.008 n=5+5)

name     old alloc/op   new alloc/op   delta
Flush-8    1.03GB ± 0%    0.20GB ± 1%  -80.64%  (p=0.016 n=4+5)

name     old allocs/op  new allocs/op  delta
Flush-8     8.00k ± 0%     5.07k ± 1%  -36.62%  (p=0.016 n=4+5)
```